I finished the function of adding new meal, user can build meal with previously added products and search them with searchview. I've also added the possibility to delete meals and improved the way products are deleted

+++ app/src/main/java/com/example/dell/menu/objects/User.java
@@ -7,13 +7,13 @@ package com.example.dell.menu.objects;
 public class User {
     private String login;
     private String password;
-    private int userId;
+    private Long userId;
 
     public User(String login, String password){
         this.login = login;
         this.password = password;
 
-        userId = -1;
+        userId = Long.valueOf(-1);
     }
 
     public String getLogin() {
@@ -24,11 +24,11 @@ public class User {
         return password;
     }
 
-    public int getUserId() {
+    public Long getUserId() {
         return userId;
     }
 
-    public void setUserId(int usersId) {
+    public void setUserId(Long usersId) {
         this.userId = usersId;
     }
 }
